refactor(team): extract state persistence from broker.go
Phase 3 fifth slice. Pulls JSON state file I/O into a kin file:

  - defaultBrokerStatePath: WUPHF_BROKER_STATE_PATH override + default
    under config.RuntimeHomeDir()
  - stateSnapshotPath: <path>.last-good
  - loadBrokerStateFile / brokerStateActivityScore /
    brokerStateShouldSnapshot: read + score the state file
  - loadState: replays the higher-activity-score of (current, snapshot)
    into broker fields; migrates legacy dm-* slugs; calls
    ensureDefaultOfficeMembersLocked + normalizeLoadedStateLocked
  - saveLocked: zero-state branch removes both files; otherwise
    atomicWriteFile current + snapshot when activity-score > 0
  - atomicWriteFile: os.CreateTemp + rename, fixes the
    concurrent-saves race documented in broker_save_race_test.go

  - internal/team/broker.go: -252 LOC (2376 -> 2124 -> 2121 after
    fmt collapses unused-import block)
  - internal/team/broker_persistence.go: +277 LOC (new)

Cumulative broker.go reduction across Phase 3:
  3404 -> 2859 (web proxy + sse)
       -> 2800 (auth + publish)
       -> 2700 (wiki lifecycle)
       -> 2376 (queries)
       -> 2121 (persistence)
  Net: -1283 LOC; broker.go shrunk by 38%.

No behavior change. Tests pass.
